requirements.txt
docx==0.2.4
Flask==3.0.0
langchain==0.0.337
PyPDF2==3.0.1
python-dotenv==1.0.0
